% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amlr_dbPool.R
\name{amlr_dbPool}
\alias{amlr_dbPool}
\title{Wrapper to connect to an AMLR database}
\usage{
amlr_dbPool(
  Database,
  Driver = "ODBC Driver 18 for SQL Server",
  Server = "swc-estrella-s",
  Trusted_Connection = "Yes",
  Encrypt = "Optional",
  idleTimeout = 3600000,
  ...
)
}
\arguments{
\item{Database}{character; database name}

\item{Driver}{character; default is 'ODBC Driver 18 for SQL Server'}

\item{Server}{character; default is 'swc-estrella-s'}

\item{Trusted_Connection}{character; default is 'Yes'}

\item{Encrypt}{character; default is 'Optional'}

\item{idleTimeout}{integer; default is one hour}

\item{...}{additional arguments passed to \link[pool]{dbPool}}
}
\value{
Output of \code{\link[base]{try}(\link[pool]{dbPool})} call
}
\description{
Wrapper with standard, AMLR defaults for connecting to an AMLR database
}
\details{
Wrapper for a call to \code{\link[pool]{dbPool}}.
See these docs for more information about these arguments
}
\examples{
\dontrun{
amlr_dbPool("AMLR_PINNIPEDS")
}

}
\seealso{
https://github.com/rstudio/pool
}
