% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amlrDatabases_shiny.R
\name{amlrDatabases_shiny}
\alias{amlrDatabases_shiny}
\title{Open the amlrDatabases Shiny app}
\usage{
amlrDatabases_shiny(...)
}
\arguments{
\item{...}{arguments passed as a list to the \code{options} argument
of \code{\link[shiny]{runApp}}}
}
\description{
Open the amlrDatabases Shiny app
}
\examples{
if (interactive()) shiny_app(launch.browser = TRUE)

}
