% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_output.R
\name{mod_output}
\alias{mod_output}
\alias{mod_output_ui}
\alias{mod_output_server}
\title{Shiny module for output plot and table for AMLR shiny apps}
\usage{
mod_output_ui(id, ...)

mod_output_server(id, tbl.reac, plot.reac, plot.res = 96)
}
\arguments{
\item{id}{module namespace, see \code{shiny::\link[shiny]{NS}}}

\item{...}{UI elements to be displayed in the plot box,
below the plot and above the save plot button}

\item{tbl.reac}{reactive; data frame to be displayed in the table}

\item{plot.reac}{reactive; \code{\link[ggplot2]{ggplot}} object to be plotted}

\item{plot.res}{numeric; plot resolution.
Value is passed to \code{\link[shiny]{renderPlot}},
and used to determine output file dimensions.}
}
\value{
Nothing
}
\description{
Shiny module for output plot and table for AMLR shiny apps
}
