% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/round_logical.R
\name{round_logical}
\alias{round_logical}
\title{Logical rounding}
\usage{
round_logical(x, digits)
}
\arguments{
\item{x}{numeric; see \code{\link[base]{round}}}

\item{digits}{numeric; see \code{\link[base]{round}}}
}
\description{
Always round fives up
}
\details{
R's default rounding function uses the IEC 60559 standard ('go to
the even digit'; see \code{\link[base]{Round:round}} for more details).
While this is useful for all of the reasons outlined in the standard,
sometimes AMLR scientists want to reproduce 'logical' or 'traditional'
rounding, meaning rounding up from 5.

This function is implemented from \url{https://stackoverflow.com/a/12688836},
and is the same as the \code{janitor}
function \code{\link[janitor]{round_half_up}} (also see web link below).
}
\seealso{
\url{https://sfirke.github.io/janitor/reference/round_half_up.html},
\url{https://appsilon.com/rounding-issues-in-r-julia-and-python/}
}
