% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amlr_dbPool.R
\name{amlr_dbPool}
\alias{amlr_dbPool}
\title{Wrapper to connect to an AMLR database}
\usage{
amlr_dbPool(
  Database,
  Driver = "SQL Server",
  Server = "swc-estrella-s",
  idleTimeout = 3600000,
  ...
)
}
\arguments{
\item{Database}{character; see \code{\link[pool]{dbPool}}}

\item{Driver}{character; default is 'SQL Server'. See \code{\link[pool]{dbPool}}}

\item{Server}{character; default is 'swc-estrella-s'. See \code{\link[pool]{dbPool}}}

\item{idleTimeout}{integer; default is one hour. See \code{\link[pool]{dbPool}}}

\item{...}{additional arguments passed to \link[pool]{dbPool})}
}
\value{
Output of \code{\link[base]{try}(\link[pool]{dbPool})} call
}
\description{
Wrapper with standard, AMLR defaults for connecting to an AMLR database
}
\details{
Wrapper for a call to \link[pool]{dbPool})
}
\examples{
\dontrun{
amlr_dbPool("AMLR_PINNIPEDS")
}

}
\seealso{
https://github.com/rstudio/pool
}
