% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amlr_season.R
\name{amlr_season}
\alias{amlr_season}
\title{Determine AMLR season from date}
\usage{
amlr_season(x)
}
\arguments{
\item{x}{object of class Date}
}
\value{
Character vector of length \code{x} of the calculated season names,
in the form 'YYYY/YY' (e.g., 2016/17)
}
\description{
Determine AMLR season from date
}
\details{
Determine the AMLR season from the date.
Note that this just generates the season name from the date year,
with July as the demarcation,
rather than using the season_info table.
}
\examples{
amlr_season(as.Date("1999-12-31"))
amlr_season(as.Date(c("2002-03-01", "2002-10-01")))
amlr_season(as.Date("2017-01-01"))

}
