% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amlrDatabases-package.R
\docType{package}
\name{amlrDatabases-package}
\alias{amlrDatabases-package}
\alias{amlrDatabases}
\title{General Functions for AMLR Database Packages}
\description{
amlrDatabases: General functions used by other AMLR packages
}
\details{
General functions used by other AMLR packages
}
\author{
Sam Woodman \email{sam.woodman@noaa.gov}
}
\keyword{package}
