% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amlrDatabases_shiny.R
\name{amlrDatabases_shiny}
\alias{amlrDatabases_shiny}
\title{Open the amlrDatabases Shiny app}
\usage{
amlrDatabases_shiny(launch.browser = TRUE)
}
\arguments{
\item{launch.browser}{Logical with default of \code{TRUE};
passed to \code{launch.browser} argument of \code{\link[shiny]{runApp}}}
}
\description{
Open the amlrDatabases Shiny app (for testing)
}
\examples{
if (interactive()) tamatoa(launch.browser = TRUE)

}
\seealso{
\url{https://www.fisheries.noaa.gov/about/antarctic-ecosystem-research-division-southwest-fisheries-science-center}
}
