% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_database.R
\name{mod_database}
\alias{mod_database}
\alias{mod_database_ui}
\alias{mod_database_server}
\title{Database module for AMLR shiny apps}
\usage{
mod_database_ui(
  id,
  db.name.prod,
  db.name.test,
  remote.valid,
  col.width = 5,
  db.remote.default = "remote_prod"
)

mod_database_server(id, pool.remote.prod, pool.remote.test, db.driver)
}
\arguments{
\item{id}{character used to specify namespace, see \code{shiny::\link[shiny]{NS}}}

\item{db.name.prod}{character; name of production database on SWFSC server}

\item{db.name.test}{character; name of test database on SWFSC server}

\item{remote.valid}{logical; indicates if the remote (TRUE) or local (FALSE)
database connection should be the default}

\item{col.width}{integer; column width of column of UI widgets}

\item{db.remote.default}{character; default remote database to connect to.
To allow developer to specify test database as initial db}

\item{pool.remote.prod}{output of a \code{\link[pool]{dbPool}} call.
A DBI database connection pool connected to the remote PRODUCTION database,
e.g. 'AMLR_PINNIPEDS'}

\item{pool.remote.test}{output of a \code{\link[pool]{dbPool}} call.
A DBI database connection pool connected to the remote TEST database,
e.g. 'AMLR_PINNIPEDS_Test'}

\item{db.driver}{character; name of driver used to connect to the databases}
}
\value{
\code{mod_database_server} returns a reactive of the pool connection
specified by the user
}
\description{
Database module for AMLR shiny apps
}
\details{
This module allows users to connect to the database of their choice:
the provided production database, the provided dev (test) database,
or a database specified by the user (e.g., a local database).

If the user specifies their own database,
then the pool object is created and returned by this module.
}
