% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_database.R
\name{mod_database}
\alias{mod_database}
\alias{mod_database_ui}
\alias{mod_database_server}
\title{Database module for AMLR shiny apps}
\usage{
mod_database_ui(
  id,
  col.width = 5,
  server.default = NULL,
  database.default = "AMLR_PINNIPEDS",
  port.check.default = FALSE,
  port.default = 1443,
  conn.default = "trusted",
  uid.default = "sa",
  pwd.default = NULL
)

mod_database_server(
  id,
  pool.list = list(),
  db.driver = "ODBC Driver 18 for SQL Server",
  db.selected = NULL
)
}
\arguments{
\item{id}{module namespace, see \code{shiny::\link[shiny]{NS}}}

\item{col.width}{integer; column width of column of UI widgets}

\item{server.default}{default character value for "SQL Server Name";
default is \code{NULL}. If \code{NULL}, default in module is
\code{paste0(\link{Sys.info}()[["nodename"]], "\\SQLEXPRESS")}}

\item{database.default}{default character value for "SQL Server Database" widget;
default is "AMLR_PINNIPEDS"}

\item{port.check.default}{default boolean value for 'Specify port number' widget;
default is \code{FALSE}}

\item{port.default}{default numeric value for "Port number";
default is 1443}

\item{conn.default}{default character value for "Connection type";
default is "trusted}

\item{uid.default}{default character value for "User"; default is "sa"}

\item{pwd.default}{default character value for "Password";
default is \code{NULL}}

\item{pool.list}{a (named) list of pool objects created by \code{\link[pool]{dbPool}}.
The names of the objects in this list will appear as
'database connection' \code{\link[shiny]{radioButtons}} options.}

\item{db.driver}{character; name of driver used to connect to the databases}

\item{db.selected}{character; the initial selected value for the
database connection widget (\code{output$db_conn_uiOut}).
Passed to \code{selected} argument of \code{\link[shiny]{radioButtons}}.
Must be \code{NULL}, or the name of a connection in \code{pool.list}}
}
\value{
Returns a reactive of the pool connection specified by the user
}
\description{
Database module for AMLR shiny apps
}
\details{
This module allows users to connect to the database of their choice:
any already-created connections
(e.g., to the AMLR_PINNIPEDS production database on the SWC server)
or a database specified by the user.

If the user specifies their own database,
then this module creates the pool object.
}
