% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_database.R
\name{mod_database}
\alias{mod_database}
\alias{mod_database_ui}
\alias{mod_database_server}
\title{Database module for AMLR shiny apps}
\usage{
mod_database_ui(id, col.width = 5)

mod_database_server(id, pool.list = list(), db.driver)
}
\arguments{
\item{id}{module namespace, see \code{shiny::\link[shiny]{NS}}}

\item{col.width}{integer; column width of column of UI widgets}

\item{pool.list}{a (named) list of pool objects created by \code{\link[pool]{dbPool}}.
The names of the objects in this list will appear as
'database connection' radioButton options.
The default is an empty list}

\item{db.driver}{character; name of driver used to connect to the databases.
Default is 'ODBC Driver 18 for SQL Server'}
}
\value{
Returns a reactive of the pool connection specified by the user
}
\description{
Database module for AMLR shiny apps
}
\details{
This module allows users to connect to the database of their choice:
any already-created connections
(e.g., to the AMLR_PINNIPEDS production database on the SWC server)
or a database specified by the user.

If the user specifies their own database,
then this module creates the pool object.
}
